update  for all the custom select
