+++ src/SharedComponents/CustomSelect/SelectBanqueComponent/index.tsx
@@ -1,11 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import Grid from '@mui/material/Grid'
-import FormControl from '@mui/material/FormControl'
 import Select from 'react-select';
-import { getOptions } from './mockAdherent';
-import {  SelectInerface } from '../SelectInterface';
+import { getOptions } from './mockBanque';
+import { CustomSelectProps, SelectInerface } from '../interface/customSelect.interface';
 
-const SelectAdherent = ({ onSearch, labelText }: any) => {
+
+
+
+
+const SelectBanque = ({ onSearch, labelText }: CustomSelectProps) => {
 
   const [options, setOptions] = useState<SelectInerface[]>([]);
   const [statusValue, setStatusValue] = useState('');
@@ -26,30 +29,24 @@ const SelectAdherent = ({ onSearch, labelText }: any) => {
 
 
   return (
-    
-      <Grid container spacing={6}>
-        <Grid item xs={12}>
+   
           <Grid container spacing={6}>
             <Grid item xs={12} sm={6}>
-              <FormControl fullWidth sx={{ mr: 4, mb: 2 }}>
 
-                <Select placeholder={'---Sélectionnez un adhérent---'}
+                <Select placeholder={'---Sélectionnez une banque---'}
                   defaultValue={statusValue}
                   onChange={(value: any) => {
-                    onSearch ? onSearch() :  setStatusValue(value)
-                    labelText 
-                   }}
+                   onSearch ? onSearch(value):  setStatusValue(value)
+                   labelText 
+                  }}
                   options={options} 
                   isClearable={isClearable}
                   />
-              </FormControl>
             </Grid>
           </Grid>
-        </Grid>
-      </Grid>
-  
+       
 
   )
 }
 
-export default SelectAdherent;
+export default SelectBanque;
