+++ src/SharedComponents/CustomSelect/index.ts
@@ -1,0 +1,7 @@
+import  SelectModePaiement  from './SelectModePaimentComponent';
+import  SelectIndividu  from './SelectIndividualComponent';
+import  SelectDevise  from './SelectDeviseComponent';
+import SelectAdherent from './SelectAdherentComponent'
+import SelectBanque from './SelectBanqueComponent'
+
+export default { SelectAdherent, SelectBanque ,SelectDevise,SelectIndividu,SelectModePaiement}
