+++ src/SharedComponents/CustomSelect/interface/customSelect.interface.ts
@@ -1,0 +1,8 @@
+export interface SelectInerface {
+    value: string
+    label: string
+  }
+  export interface CustomSelectProps {
+    onSearch: (value: any) => void;
+    labelText: string;
+  }
