--- src/SharedComponents/SelectInterface.ts
@@ -1,7 +1,0 @@
-export interface SelectInerface{
-        value: string;
-        label: string;
-         
-}
-
-
